SurfaceReactions

Documentation for SurfaceReactions.

SurfaceReactions.calculate_molar_production_rates!

SurfaceReactions.calculate_ss_molar_production_rates!

SurfaceReactions.compile_mech

SurfaceReactions.create_mech

SurfaceReactions.update_params

SurfaceReactions.update_params!

SurfaceReactions.calculate_molar_production_rates! — Method

calculatemolarproductionrates!(state::ReactionState, thermoobj::IdealGas.SpeciesThermoObj,md::SurfaceMechDefinition)

Function to calculate the molar production rate of gas-phase and surface species based on the surface reaction mechanism. The function will alter the field source after the calculation.

Usage:

calculate_molar_production_rate!(state,md)

state::State : state of the reaction, which is of the composite type State

thermo_obj::SpeciesThermoObj : Composite type SpeciesThermoObj. Required in case of reversible reactions

md::MechanismDefinition : the mechanism, which is of the composite type MechanismDefinition

SurfaceReactions.calculate_ss_molar_production_rates! — Function

calculatessmolarproductionrates!(state::ReactionState, thermo_obj::IdealGas.SpeciesThermoObj,md::SurfaceMechDefinition, time =1.0)

Function to calculate the molar production rate of gas-phase species. This function integrates the rate equations so that rac{d heta}{dt}=0

Usage:

calculate_molar_production_rate!(state,md)

state::State : state of the reaction, which is of the composite type State

thermo_obj::SpeciesThermoObj : Composite type SpeciesThermoObj. Required in case of reversible reactions

md::MechanismDefinition : the mechanism, which is of the composite type MechanismDefinition

SurfaceReactions.compile_mech — Method

Function to create composite type of SurfaceMechanism and SpeciesRxnMap. In general the application programs shall only call the compilemech function and not directly the createmech and speciesrxnmap function

Usage

compile_mech(file_path,gas_species)

file_path::String : path to the mechanism xml file

gas_species::Array{String,1} : list of gasphase species

SurfaceReactions.create_mech — Method

Function to read a surface reaction mechanism file

Usage:

create_mech(file_path,gas_species)

file_path::String : file name including the relative path

gas_species::Array{String,1} : gasphase species list

SurfaceReactions.update_params! — Method

A function for use in global sensitivity analysis. The function updates the parameters with the supplied rate_params

Usage

update_params!(md, rate_params,rxn_ids, constraint_ids)

md : SurfaceMechDefinition structure

rate_params : parameters to that needs to be set to the mechanism object

rxn_ids : id of the reactions considered in the sensitivity analysis

constraint_ids : id of the reverse reaction ids considered

SurfaceReactions.update_params — Method

Function to update the rate parameters. This function is only used for sensitivity analysis

Usage

update_params(rate_params,md)

rate_params::Array : One dimentional array of floating point variables

md:SurfaceMechDefinition : SurfaceMechDefinition struct